modify hydra
