+++ examples/topopt/functions.py
@@ -12,8 +12,70 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import h5py
 import numpy as np
 import paddle
+
+
+def uniform_sampler() -> callable:
+    return lambda: np.random.randint(1, 99)
+
+
+def poisson_sampler(lam: int) -> callable:
+    def func():
+        iter_ = max(np.random.poisson(lam), 1)
+        iter_ = min(iter_, 99)
+        return iter_
+
+    return func
+
+
+def generate_sampler(sampler_type: str = "Fixed", num: int = 0) -> callable:
+    """Generate sampler for the number of initial iteration steps
+
+    Args:
+        sampler_type (str): "Poisson" for poisson sampler; "Uniform" for uniform sampler; "Fixed" for choosing a fixed number of initial iteration steps.
+        num (int): If `sampler_type` == "Poisson", `num` specifies the poisson rate parameter; If `sampler_type` == "Fixed", `num` specifies the fixed number of initial iteration steps.
+
+    Returns:
+        sampler (callable): sampler for the number of initial iteration steps
+    """
+    if sampler_type == "Poisson":
+        return poisson_sampler(num)
+    elif sampler_type == "Uniform":
+        return uniform_sampler()
+    else:
+        return lambda: num
+
+
+def generate_train_test(data_path, train_test_ratio, n_sample):
+    h5data = h5py.File(data_path, "r")
+    X = h5data["iters"]
+    Y = h5data["targets"]
+    idx = np.arange(10000)
+    np.random.shuffle(idx)
+    train_idx = idx <= train_test_ratio * n_sample
+    if train_test_ratio == 1:
+        X_train = []
+        Y_train = []
+        for i in range(10000):
+            if train_idx[i]:
+                X_train.append(np.array(X[i]))
+                Y_train.append(np.array(Y[i]))
+        return X_train, Y_train
+    else:
+        X_train = []
+        X_test = []
+        Y_train = []
+        Y_test = []
+        for i in range(10000):
+            if train_idx[i]:
+                X_train.append(np.array(X[i]))
+                Y_train.append(np.array(Y[i]))
+            else:
+                X_test.append(np.array(X[i]))
+                Y_test.append(np.array(Y[i]))
+        return X_train, Y_train, X_test, Y_test
 
 
 def augmentation(input_dict, label_dict, weight_dict=None):
