+++ examples/topopt/topopt.py
@@ -14,43 +14,139 @@
 
 import os
 
+import hydra
 import numpy as np
 import paddle
-from data_utils import augmentation
+from functions import augmentation
+from functions import generate_sampler
+from functions import generate_train_test
 from matplotlib import pyplot as plt
+from omegaconf import DictConfig
 from paddle import nn
-from prepare_datasets import generate_train_test
-from sampler_utils import generate_sampler
 from TopOptModel import TopOptNN
 
 import ppsci
-from ppsci.utils import config
 from ppsci.utils import logger
 
 
-def evaluation_and_plot():
-    args = config.parse_args()
-    OUTPUT_DIR = "Output_TopOpt" if args.output_dir is None else args.output_dir
-    RESULT_DIR = os.path.join(OUTPUT_DIR, "results")
-    DATA_PATH = "./Dataset/PreparedData/top_dataset.h5"
-    # initialize logger
-    logger.init_logger("ppsci", f"{RESULT_DIR}/results.log", "info")
+def train(cfg: DictConfig):
+    ppsci.utils.misc.set_random_seed(cfg.seed)
 
-    # specify evaluation parameters
-    BATCH_SIZE = 32
-    VOL_COEFF = 1  # coefficient for volume fraction constraint in the loss - beta in equation (3) in paper
-    NUM_VAL_STEP = 10  # the number of iteration for each evaluation case
-    MODEL_LIST = ["Poisson5", "Poisson10", "Poisson30", "Uniform"]
+    # 4 training cases parameters
+    LEARNING_RATE = 0.001 / (1 + cfg.TRAIN.num_epochs // 15)
+    ITERS_PER_EPOCH = int(
+        cfg.TRAIN.n_samples * cfg.TRAIN.train_test_ratio / cfg.TRAIN.batch_size
+    )
+
+    # generate training dataset
+    X_train, Y_train = generate_train_test(
+        cfg.DATA_PATH, cfg.TRAIN.train_test_ratio, cfg.TRAIN.n_samples
+    )
 
     # define loss
     def loss_expr(output_dict, label_dict, weight_dict=None):
         y = label_dict["output"].reshape((-1, 1))
         y_pred = output_dict["output"].reshape((-1, 1))
         conf_loss = paddle.mean(
-            nn.functional.log_loss(y_pred, y, epsilon=1e-7)
-        )  # epsilon = 1e-07 is the default in tf
+            nn.functional.log_loss(y_pred, y, epsilon=cfg.TRAIN.epsilon.log_loss)
+        )
         vol_loss = paddle.square(paddle.mean(y - y_pred))
-        return conf_loss + VOL_COEFF * vol_loss
+        return conf_loss + cfg.TRAIN.vol_coeff * vol_loss
+
+    # set constraints
+    sup_constraint = ppsci.constraint.SupervisedConstraint(
+        {
+            "dataset": {
+                "name": "NamedArrayDataset",
+                "input": {"input": X_train},
+                "label": {"output": Y_train},
+            },
+            "batch_size": cfg.TRAIN.batch_size,
+            "sampler": {
+                "name": "BatchSampler",
+                "drop_last": cfg.TRAIN.drop_last,
+                "shuffle": cfg.TRAIN.shuffle,
+            },
+            "transforms": (
+                {
+                    "FunctionalTransform": {
+                        "transform_func": augmentation,
+                    },
+                },
+            ),
+        },
+        ppsci.loss.FunctionalLoss(loss_expr),
+        name="sup_constraint",
+    )
+    constraint = {sup_constraint.name: sup_constraint}
+
+    # train models for 4 cases
+    for sampler_key, num in cfg.CASE_PARAM:
+
+        # initialize SIMP iteration stop time sampler
+        SIMP_stop_point_sampler = generate_sampler(sampler_key, num)
+
+        # initialize logger for training
+        OUTPUT_DIR = cfg.output_dir
+        OUTPUT_DIR = (
+            f"{OUTPUT_DIR}/{sampler_key}{num}_vol_coeff{cfg.TRAIN.vol_coeff}"
+            if num is not None
+            else f"{OUTPUT_DIR}/{sampler_key}_vol_coeff{cfg.TRAIN.vol_coeff}"
+        )
+        logger.init_logger("ppsci", os.path.join(OUTPUT_DIR, "train.log"), "info")
+
+        # set model
+        model = TopOptNN(
+            in_channel=cfg.MODEL.in_channel,
+            out_channel=cfg.MODEL.out_channel,
+            kernel_size=cfg.MODEL.kernel_size,
+            filters=cfg.MODEL.filters,
+            layers=cfg.MODEL.layers,
+            channel_sampler=SIMP_stop_point_sampler,
+        )
+        assert model.num_params == cfg.TRAIN.num_params
+
+        # set optimizer
+        optimizer = ppsci.optimizer.Adam(
+            learning_rate=LEARNING_RATE, epsilon=cfg.TRAIN.epsilon.optimizer
+        )(model)
+
+        # initialize solver
+        solver = ppsci.solver.Solver(
+            model,
+            constraint,
+            OUTPUT_DIR,
+            optimizer,
+            epochs=cfg.TRAIN.num_epochs,
+            iters_per_epoch=ITERS_PER_EPOCH,
+        )
+
+        # train model
+        solver.train()
+
+
+def evaluate_and_plot(cfg: DictConfig):
+    ppsci.utils.misc.set_random_seed(cfg.seed)
+
+    # evaluate 4 models
+    RESULT_DIR = os.path.join(cfg.output_dir, "results")
+
+    # initialize logger for evaluation
+    logger.init_logger("ppsci", os.path.join(RESULT_DIR, "results.log"), "info")
+
+    # fixed iteration stop times for evaluation
+    iterations_stop_times = range(5, 85, 5)
+    model = TopOptNN()
+
+    # define loss
+    def loss_expr(output_dict, label_dict, weight_dict=None):
+        y = label_dict["output"].reshape((-1, 1))
+        y_pred = output_dict["output"].reshape((-1, 1))
+        conf_loss = paddle.mean(
+            nn.functional.log_loss(y_pred, y, epsilon=cfg.TRAIN.epsilon.log_loss)
+        )
+        vol_loss = paddle.square(paddle.mean(y - y_pred))
+        return conf_loss + cfg.TRAIN.vol_coeff * vol_loss
 
     # define metric
     def val_metric(output_dict, label_dict, weight_dict=None):
@@ -96,22 +192,16 @@ def evaluation_and_plot():
         current_iou_results.append(np.array(iou))
         return {"Binary_Acc": acc, "IoU": iou}
 
-    # fixed iteration stop times for evaluation
-    iterations_stop_times = range(5, 85, 5)
-    model = TopOptNN()
-
     # evaluation for 4 cases
     acc_results_summary = {}
     iou_results_summary = {}
-    for model_name in MODEL_LIST:
+    for model_name in cfg.EVAL.model_list:
 
         # load model parameters
-        model_path = (
-            f"{OUTPUT_DIR}/{model_name}_vol_coeff{VOL_COEFF}/checkpoints/latest"
-        )
+        model_path = f"{cfg.output_dir}/{model_name}_vol_coeff{cfg.TRAIN.vol_coeff}/checkpoints/latest"
         solver = ppsci.solver.Solver(model, pretrained_model_path=model_path)
         solver.epochs = 1
-        solver.iters_per_epoch = NUM_VAL_STEP
+        solver.iters_per_epoch = cfg.EVAL.num_val_step
         acc_results = []
         iou_results = []
 
@@ -119,7 +209,7 @@ def evaluation_and_plot():
         for stop_iter in iterations_stop_times:
             # only evaluate for NUM_VAL_STEP times of iteration
             X_data, Y_data = generate_train_test(
-                DATA_PATH, 1, BATCH_SIZE * NUM_VAL_STEP
+                cfg.DATA_PATH, 1, cfg.EVAL.batch_size * cfg.EVAL.num_val_step
             )
             sup_validator = ppsci.validate.SupervisedValidator(
                 {
@@ -128,11 +218,11 @@ def evaluation_and_plot():
                         "input": {"input": X_data},
                         "label": {"output": Y_data},
                     },
-                    "batch_size": BATCH_SIZE,
+                    "batch_size": cfg.EVAL.batch_size,
                     "sampler": {
                         "name": "BatchSampler",
-                        "drop_last": False,
-                        "shuffle": True,
+                        "drop_last": cfg.EVAL.drop_last,
+                        "shuffle": cfg.EVAL.shuffle,
                     },
                     "transforms": (
                         {
@@ -175,7 +265,7 @@ def evaluation_and_plot():
 
         # only calculate for NUM_VAL_STEP times of iteration
         for i in range(10):
-            x, y = generate_train_test(DATA_PATH, 1, BATCH_SIZE)
+            x, y = generate_train_test(cfg.DATA_PATH, 1, cfg.EVAL.batch_size)
             # thresholding
             k = SIMP_stop_point_sampler()
             x1 = paddle.to_tensor(x)[:, k, :, :]
@@ -212,3 +302,17 @@ def evaluation_and_plot():
     plt.grid()
     plt.savefig(os.path.join(RESULT_DIR, "IoU.png"))
     plt.show()
+
+
+@hydra.main(version_base=None, config_path="./conf", config_name="topopt.yaml")
+def main(cfg: DictConfig):
+    if cfg.mode == "train":
+        train(cfg)
+    elif cfg.mode == "eval":
+        evaluate_and_plot(cfg)
+    else:
+        raise ValueError(f"cfg.mode should in ['train', 'eval'], but got '{cfg.mode}'")
+
+
+if __name__ == "__main__":
+    main()
